requirements.txt
selenium
opencv-python-headless
openai
psycopg2-binary
python-dotenv
